Move Accumulator to yash-env
