+++ yash-env/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `yash-env` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.1] - Unreleased
+
+### Added
+
+- `job::fmt::Accumulator`
+    - This is a utility for formatting status reports of multiple jobs.
+
 ## [0.3.0] - 2024-08-22
 
 ### Added
@@ -261,6 +268,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of the `yash-env` crate
 
+[0.3.1]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.3.1
 [0.3.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.3.0
 [0.2.1]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.2.1
 [0.2.0]: https://github.com/magicant/yash-rs/releases/tag/yash-env-0.2.0
